+++ src/utils/googleAnalytics.js
@@ -81,9 +81,6 @@ export function normalizeGoogleAnalyticsRoute(value) {
 
   if (STATIC_ROUTE_TEMPLATES.has(pathname)) return pathname;
   if (/^\/ingredients\/[^/]+\/edit$/u.test(pathname)) return '/ingredients/:id/edit';
-  if (/^\/recipes\/ingredients\/[^/]+$/u.test(pathname)) return '/recipes/ingredients/:slug';
-  if (/^\/recipes\/[^/]+$/u.test(pathname)) return '/recipes/:slug';
-  if (/^\/guides\/[^/]+$/u.test(pathname)) return '/guides/:slug';
   return '/other';
 }